requirements.txt
pdfplumber>=0.11.0
python-docx>=0.8.11
click>=8.1.0
PyPDF2>=3.0.0
